+++ CHANGES.md
@@ -1,3 +1,9 @@
+## [0.6] - 2018-05-31
+Added documentation on Read the Docs.
+### Deprecated
+ - FakeMetrics is now a list. Fakemetrics.metrics will be removed in 1.0
+   release.
+
 ## [0.5] - 2018-04-27
 ### Breaking changes
  - Gauges no longer record min/mean/max/count
@@ -5,8 +11,8 @@
    for most gauges, the min/max/mean params aren't helpful.
 
    We've decided to drop that feature from `recordGauge` replacing it with a 
-   new `recordHistogram` method. The `time` method has been rewritten to use a
-   histogram rather than a gauge.
+   new `recordSummary` method. The `time` method has been rewritten to use a
+   Summary rather than a Gauge.
  - Made the "counters", and "gauges" properties of the Metrics class private.
  - Made the state of Counter, and Gauge private.
 
@@ -14,6 +20,7 @@
 ## [0.4] - 2017-11-14
 ### Fixes
 -  Fix issue where we get stuck always 'Failed to flush metrics to riemann'
+
 ### Deprecated
 - RiemannTransport.is_connected() should no longer be needed
 
@@ -30,5 +37,6 @@
 - TTL attributes are no longer coerced to strings
 
 
+[0.6]: https://github.com/madedotcom/striemann/compare/v0.5...v0.6
 [0.5]: https://github.com/madedotcom/striemann/compare/v0.4.0...v0.5
 [0.4]: https://github.com/madedotcom/striemann/compare/v0.3.0...v0.4.0